+++ testdata/apps/custom_main/app/setup.go
@@ -1,0 +1,6 @@
+package app
+
+import "github.com/emersonjoe/trilha"
+
+func Setup(a *trilha.App) error    { return nil }
+func Shutdown(a *trilha.App) error { return nil }