+++ testdata/apps/dotdir/app/manifest.webmanifest/route.go
@@ -1,0 +1,5 @@
+package manifest
+
+import "github.com/emersonjoe/trilha"
+
+func GET(c *trilha.Ctx) error { return c.JSON(200, map[string]string{"name": "x"}) }